Atualizacoes e melhoras nos models

+++ src/api/model/db.js
@@ -10,38 +10,31 @@ const sequelize = new Sequelize(config.server.dbName, config.server.user, config
 const start = async () => {
   try {
     await sequelize.authenticate();
-    console.log('Connection has been established successfully.');
 
     //Depois de criar as tabelas, comente esse trecho
     for (const key in db) {
       if (db[key].sync) {
-        await db[key].sync({force: true});
-        console.log(`Table for ${key} synchronized successfully.`);
+        const model = db[key];
+        const tableExists = await model.sync({force: false});
+
+        if(!tableExists){
+          await db[key].sync({force: true});
+        }
+        console.log(`Tabela para ${key} sincronizada.`);
       }
     }
-
-    //Descomente depois de criar as tabelas
-
-    /*
-    for (const key in db) {
-      if (db[key].sync) {
-        await db[key].sync();
-        console.log(`Table for ${key} synchronized successfully.`);
-      }
-    }
-    */
 
   } catch (error) {
-    console.error('Unable to connect to the database:', error);
+    console.error('Falha na conexão: ', error);
   }
 };
 
 const stop = async () => {
   try {
     await sequelize.close();
-    console.log('Connection to the database has been closed.');
+    console.log('Conexão fechada.');
   } catch (error) {
-    console.error('Error closing connection:', error);
+    console.error('Erro ao fechar conexão:', error);
   }
 };
